Add dedicated class for the test kit
